+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All Notable changes to `bakame/aide-profiler` will be documented in this file.
 - `Label::random` accepts an optional `$length` parameter to adjust the label length.
 - `Marker::complete` and `Marker::isComplete`
 - `Profiler::hasSummaries`
+- `Snapshot::fromArray`
+- `Metrics::fromArray`
+- `Summary::fromArray`
 
 ### Fixed
 
@@ -19,6 +22,7 @@ All Notable changes to `bakame/aide-profiler` will be documented in this file.
 - **BC BREAK:** `Exporter::exportProfilingData` is renamed `Exporter::exportSummary`
 - **BC BREAK:** `Snapshot::toArray` and `Snapshot::jsonSerialize` representation simplified
 - **BC BREAK:** `Summary::toArray` and `Summary::jsonSerialize` representation simplified
+- **BC BREAK:** `Summary::__construct`signature changed
 - **BC BREAK:** `Profiler::toArray` and `Profiler::jsonSerialize` representation simplified
 - **BC BREAK:** `Marker::toArray` and `Marker::jsonSerialize` representation simplified
 - **BC BREAK:** `Marker::finish` throws an `UnableToProfile` exception previously was an `InvalidArgument` exception when it cannot complete its task.